feat: balance refactoring v2
