+++ app/shared/api/blockchain/transfer/constants.ts
@@ -1,0 +1,20 @@
+import type { SignerOptions } from '@polkadot/api/types';
+
+export const ASSET_LOCATION: Record<string, SignerOptions> = {
+  // USDT MultiLocation
+  1984: {
+    // @ts-expect-error type error
+    parents: 0,
+    interior: {
+      X2: [{ PalletInstance: 50 }, { GeneralIndex: 1984 }],
+    },
+  },
+  // DOT MultiLocation
+  0: {
+    // @ts-expect-error type error
+    parents: 1,
+    interior: {
+      Here: '',
+    },
+  },
+};
