+++ app/shared/api/blockchain/transfer/types.ts
@@ -6,7 +6,6 @@ export interface ITransfer {
   sendTransfer: (params: SendTransferParams) => Promise<Hash>;
   getTransferFee: (params: FeeParams) => Promise<BN>;
   getGiftTransferFee: (params: FeeParams) => Promise<BN>;
-  getExistentialDeposit: () => Promise<BN>;
 }
 
 export type SendTransferParams = {
