+++ app/shared/api/types.ts
@@ -1,2 +1,3 @@
 export { type ProviderWithMetadata } from './blockchain/network/provider-api';
 export { type ITransfer } from './blockchain/transfer/types';
+export { type IBalance } from './blockchain/balances/types';
