+++ app/shared/helpers/constants.ts
@@ -1,5 +1,3 @@
-import { type SignerOptions } from '@polkadot/api/types';
-
 // CloudStorage / LocalStorage keys
 export const PUBLIC_KEY_STORE = 'publicKey';
 export const MNEMONIC_STORE = 'mnemonic';
@@ -10,23 +8,3 @@ export const BACKUP_DATE = 'backupDate';
 export const FAKE_ACCOUNT_ID = '0x' + '1'.repeat(64);
 
 export const SS58_DEFAULT_PREFIX = 42;
-
-// TODO: take json from nova-helpers
-export const ASSET_LOCATION: Record<string, SignerOptions> = {
-  // USDT MultiLocation
-  1984: {
-    // @ts-expect-error type error
-    parents: 0,
-    interior: {
-      X2: [{ PalletInstance: 50 }, { GeneralIndex: 1984 }],
-    },
-  },
-  // DOT MultiLocation
-  0: {
-    // @ts-expect-error type error
-    parents: 1,
-    interior: {
-      Here: '',
-    },
-  },
-};
